requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
motor==3.3.2
pymongo==4.6.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
aiofiles==23.2.1
pydantic==2.5.2
